pricesApi > api rename
