+++ src/custom/api/gnosis/errors/MetadataError.ts
