+++ src/custom/api/gnosis/errors/OperatorError.ts
