+++ src/custom/api/gnosis/errors/QuoteError.ts
