+++ src/custom/api/gnosis/operatorMock.ts
