+++ src/custom/utils/explorer.ts
@@ -1,5 +1,5 @@
 import { SupportedChainId as ChainId } from 'constants/chains'
-import { OrderID } from 'priceApi/gnosis'
+import { OrderID } from '@src/custom/api/gnosis'
 import { isLocal, isDev, isPr, isStaging, isBarn } from './environments'
 
 function _getExplorerUrlByEnvironment() {
